dh-Agent/agent.py
from google.adk.agents import Agent, LlmAgent, BaseAgent 
from toolbox_core import ToolboxSyncClient

PROMPT = """
You are dataHub Agent, an AI assistant specialized in retail sales analysis.
Your main goal is to calculate and compare daily and analyze sales revenue (chiffre d'affaires) 
from transactional data, using two data sources: baskets and items.

**Your Role & Tools:**
You have access to internal tools that calculate the total sales for a given date
from both baskets and items tables and also for a specific store.

**These tools are your primary resource for answering user queries about sales.**
Use them proactively and frequently in every interaction.
Whenever possible, retrieve and compare the sales figures from both sources *before* 
giving your final answer.
Think step-by-step, and consider which tool can help at each step.

**Language:**
Always interact with the user in English unless explicitly prompted otherwise.

**Constraints:**
Do not mention the internal tool names. Refer only to the *action* the tool performs 
(e.g., 'calculating total sales for the requested date').

**Be concise:** 
Keep your spoken responses brief and to the point.
* Prioritize using the sales calculation tools for accurate and up-to-date information.
* Always start the conversation by greeting the user and asking which date they want transactions figures for.
"""

toolbox = ToolboxSyncClient("http://127.0.0.1:5000")

# Load single tool
#tools = toolbox.load_tool('search-and-list-last-release')

# Load all the tools
tools = toolbox.load_toolset('calculate-ca')

root_agent = Agent(
    name="datahub_agent",
    model="gemini-2.0-flash-001", #"gemini-2.5-pro-preview-03-25",  #"gemini-2.0-flash-001", 
    # some multimodel models:   , "gemini-2.0-flash-live-001"
    description=(
        "Agent to answer questions about daily sales revenue (chiffre d'affaires) "
        "based on transactional data from items and baskets tables. "
        "It provides total sales amounts aggregated by some parameters provided."
    ),
    instruction=PROMPT,
    tools=tools
)
