requirements.txt
google-cloud-bigquery
python-dotenv
# Add other dependencies as needed for your project
